login/requirements.txt
requests
mysql-connector-python
pytz